requirements.txt
about-time==4.2.1
absl-py==2.1.0
alembic==1.14.0
asttokens==3.0.0
attrs==24.3.0
autograd==1.7.0
babel==2.16.0
beautifulsoup4==4.13.3
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.0
cma==3.2.2
colorama==0.4.6
colorlog==6.9.0
colorspacious==1.1.2
comm==0.2.2
contourpy==1.2.1
cuda-python==12.6.0
cudf-cu12==24.12.0
cupy-cuda12x==13.3.0
cycler==0.12.1
dcor==0.6
debugpy==1.8.11
decorator==5.1.1
dill==0.3.9
e2cnn==0.2.1
einops==0.8.0
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.21.1
fastrlock==0.8.3
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.12.0
future==1.0.0
gdown==5.2.0
glasbey==0.2.0
grapheme==0.6.0
greenlet==3.1.1
grpcio==1.69.0
huggingface-hub==0.27.1
hyperopt==0.2.7
idna==3.10
imageio==2.36.1
imbalanced-learn==0.10.0
iniconfig==2.0.0
interface-meta==1.3.0
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
kornia==0.8.0
kornia_rs==0.1.8
liac-arff==2.5.0
libcudf-cu12==24.12.0
libkvikio-cu11==24.12.1
libkvikio-cu12==24.12.1
lightgbm==4.5.0
llvmlite==0.43.0
Mako==1.3.8
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.6.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mergedeep==1.3.4
mkdocs-get-deps==0.2.0
mkdocs-material-extensions==1.3.1
mpmath==1.3.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.60.0
numba-cuda==0.0.17.1
numpy==1.22.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvcomp-cu11==4.1.0.6
nvidia-nvcomp-cu12==4.1.0.6
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
nvtx==0.2.10
opencv-python==4.10.0.84
opt_einsum==3.4.0
optuna==4.1.0
optuna-integration==4.1.0
packaging==24.2
paginate==0.5.7
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
patsy==1.0.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
prompt_toolkit==3.0.48
protobuf==5.29.3
psutil==6.1.1
ptxcompiler-cu11==0.8.1.post2
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.8
pyarrow==18.1.0
pycparser==2.22
pycryptodome==3.21.0
Pygments==2.18.0
pylibcudf-cu12==24.12.0
pymdown-extensions==10.7.1
pynndescent==0.5.13
pynvjitlink-cu12==0.4.0
pyparsing==3.2.1
pyro-api==0.1.2
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyWavelets==1.6.0
PyYAML==6.0.2
pyyaml_env_tag==0.1
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rmm-cu12==24.12.1
rpds-py==0.22.3
scikit-image==0.19.3
scikit-learn==1.2.2
scipy==1.13.1
seaborn==0.13.2
shap==0.46.0
shapely==2.0.6
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
soupsieve==2.6
SQLAlchemy==2.0.36
stack-data==0.6.3
sympy==1.13.1
tabpfn==2.0.2
-e git+https://github.com/priorlabs/tabpfn-extensions.git@6102c53b9657b485aa270c331550ea7cbfd26408#egg=tabpfn_extensions
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2024.12.12
tomli==2.2.1
torch==2.5.1
torchgeometry==0.1.2
torchvision==0.20.1
tornado==6.4.2
tqdm==4.61.2
traitlets==5.14.3
triton==3.1.0
typeguard==4.4.1
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.2
urllib3==2.3.0
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.1
xgboost==2.1.3
xmltodict==0.14.2
